src/transcription/requirements.txt
# Whisper Transcription with VAD Optimization
# Version: 1.0.0

# Core dependencies
faster-whisper==1.2.0
torch==2.9.0
torchaudio==2.9.0

# VAD
# Note: silero-vad will be installed separately or use built-in faster-whisper VAD

# Audio processing
pydub==0.25.1
soundfile==0.12.1

# Utilities
numpy>=1.24.0
typing-extensions>=4.8.0

# Optional: Speaker diarization
pyannote.audio==3.1.1
speechbrain==1.0.3

# Development
pytest==7.4.0
pytest-asyncio==0.21.0
black==23.7.0
